+++ src/BitColorMap/BitColorMap.h
@@ -13,15 +13,15 @@ public:
 
     struct LaneData
     {
-        QVector<int> data;
-        double       angle;
+        QVector<double> data;
+        double          angle;
     };
 
     int data2rgb(int data, int *r, int *g, int *b);
-    int setData(QVector<int> &line, double angle);
+    int setData(QVector<double> &line, double angle);
 
     void drawHalfCircle(QImage *img);
-    void drawLineWithAngle(QImage *img, QVector<int> &data, double angle);
+    void drawLineWithAngle(QImage *img, QVector<double> &data, double angle);
 
     // 用于清除已经画了的线
     void drawLineColorWithAngle(QImage *img, QColor color, double angle);
